# docs/api.yml
openapi: 3.0.0
info:
  title: MOCBOT API
  description: |-
    This document outlines the specifications behind the MOCBOT API

    Some useful links:
    - The [MOCBOT API](https://github.com/MasterOfCubesAU/mocbot-api) repository
  contact:
    email: masterofcubesau@gmail.com
  version: 1.4.0
...
servers:
  - url: https://api.masterofcubesau.com/v1
...
paths:
...
  /lobby/{guild_id}:
    post:
      tags:
        - Lobby Guilds
      summary: Creates a new lobby in the guild
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild to create a new lobby in
      requestBody:
        description: The data to set for this lobby
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuildLobbyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildLobby"
        "400":
          description: Missing data for lobby/no data provided
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "409":
          description: If the user is already a leader of a lobby in that guild
  /lobby/{guild_id}/{leader_id}:
    get:
      tags:
        - Lobby Guilds
      summary: Gets lobby data from a given guild ID and leader ID
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
      responses:
        "200":
          description: OK. Returns the data of the lobby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildLobby"
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
    put:
      tags:
        - Lobby Guilds
      summary: Replaces existing lobby data with new data for a given guild ID and leader ID
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
      requestBody:
        description: The data to set for this lobby
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuildLobbyInput"
      responses:
        "200":
          description: OK. Returns the new data for the lobby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildLobby"
        "400":
          description: New data was not provided/data missing
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
    patch:
      tags:
        - Lobby Guilds
      summary: Updates existing lobby data with the new provided data for a given guild ID and leader ID
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
      requestBody:
        description: The data to update for this lobby
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuildLobbyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildLobby"
        "400":
          description: New data was not provided
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
    delete:
      tags:
        - Lobby Guilds
      summary: Deletes existing lobby data with the given guild ID and leader ID
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyObject"
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
  /lobby/{guild_id}/{leader_id}/users:
    get:
      tags:
        - Lobby Users
      summary: Gets users in a given lobby
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
      responses:
        "200":
          description: OK. Returns the users in the lobby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyUsers"
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
    post:
      tags:
        - Lobby Users
      summary: Adds a new user(s) into a given lobby
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
      requestBody:
        description: The user ID(s) to add into the lobby
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LobbyUsers"
      responses:
        "200":
          description: OK. Returns the new users added into the lobby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LobbyUsers"
        "400":
          description: No new data was provided
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
    delete:
      tags:
        - Lobby Users
      summary: Deletes a user from a given lobby
      parameters:
        - name: guild_id
          in: path
          schema:
            type: integer
          required: true
          description: The guild ID of the guild that the lobby belongs to 
        - name: leader_id
          in: path
          schema:
            type: integer
          required: true
          description: The leader ID of the leader that the lobby belongs to 
        - name: user_id
          in: path
          schema:
            type: integer
          required: true
          description: The user ID of the user to remove
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyObject"
        "400":
          description: User to delete was not provided
        "401":
          $ref: "#/components/responses/NotAuthorized"
        "404":
          description: Lobby not found
components:
  responses:
    NotAuthorized:
      description: You are not authorized to access this resource
  securitySchemes:
    APIKey:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    EmptyObject:
      type: object
...
    LobbyUsers:
      type: array
      items:
        type: string
        example: "1"
    GuildLobby:
      type: object
      properties:
        GuildID:
          type: string
          description: The guild ID that the lobby belongs to 
          example: "123"
        LobbyName:
          type: string
          description: The name of the lobby
          example: "ABC"
        VoiceChannelID:
          type: string
          description: The ID of the voice channel of the lobby
          example: "123"
        TextChannelID:
          type: string
          description: The ID of the text channel of the lobby
          example: "123"
        RoleID:
          type: string
          description: The ID of the role associated with the lobby
          example: "123"
        LeaderID:
          type: string
          description: The user ID of the leader of the lobby
          example: "123"
        InviteOnly:
          type: boolean
          description: If the lobby is invite-only
          example: true
    GuildLobbyInput:
      type: object
      properties:
        LobbyName:
          type: string
          description: The name of the lobby
          example: "ABC"
        VoiceChannelID:
          type: string
          description: The ID of the voice channel of the lobby
          example: "123"
        TextChannelID:
          type: string
          description: The ID of the text channel of the lobby
          example: "123"
        RoleID:
          type: string
          description: The ID of the role associated with the lobby
          example: "123"
        LeaderID:
          type: string
          description: The user ID of the leader of the lobby
          example: "123"
        InviteOnly:
          type: boolean
          description: If the lobby is invite-only
          example: true
...
security:
  - APIKey: []